Finished  memset() function and tests.

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: okuilboe <okuilboe@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/04/25 12:29:15 by okuilboe      #+#    #+#                 */
-/*   Updated: 2025/05/01 20:47:24 by okuilboe      ########   odam.nl         */
+/*   Updated: 2025/05/04 15:13:16 by okuilboe      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int		ft_strncmp(const char *s1, const char *s2, size_t n);
 char	*ft_strnstr(const char *big, const char *little, size_t len);
 size_t	ft_strlcpy(char *dst, const char *src, size_t size);
 size_t	ft_strlcat(char *dst, const char *src, size_t size);
-void 	*ft_memset(void *s, int c, unsigned int n);
+void 	*ft_memset(void *s, int c, size_t n);
 
 /*
 ft_memset
